convert: don't include slp in singlefile root
this is required until the buildsystem detects that slp is
provided by a .py file as well.

otherwise, it is calculated that the slp...so file is required for
creating all the .so files, which is a dependency loop.

this loop doesn't exist as the slp...so is just a compiled version of
slp.py.