fixes documentation issue #511
If a parent directory of openage contains a space, 'make doc' (or better, doxygen)
splits the absolute path from the INPUT tag in bin/Doxygen in two directorys.
That way, the necessary files cannot be found and no documentation is created.

This commit adds quotation marks to the responsible cmake file.

	geändert:               buildsystem/doxygen.cmake