doc: Update DRS file format.
Internally, Age of Empires does something like this when reading files
from a DRS archive:

```c
read_drs_file('bina', 50500);
```

Note how it uses a C char with multiple bytes. This is written out as a
little-endian integer, so the 'a' byte ends up at the lowest address,
and the 'b' byte ends up at the highest address.