codecompliance: ignore merge-commit-dates
When a PR gets merged in another year than the commits are dated,
the expected copyright year in the changed files changes to the
merge-commits year. This makes the copyright year wrong according to our
codecompliance check.
This might happen, because of a PR gets merged in a year after the last
change, or because the PR got rebased without '--ignore-date'.