buildsystem: register generated files as byproducts
This means that these generated files are not guaranteed to exist/update.
Which is good, because then cmake will not run the codegen again and again
to try to update the files. They are not updated if their content does not change,
because that would cause project rebuilds even though there's no reason
for that rebuild (the file would just be touched).