Try to use Steam path.
The logic here is that there might be a Wine bug preventing the game
from fully installing (and thus having the registry key installed), but
the assets are there. This was the case for me and thought it might not
be that rare and worth trying to work around.

The actual Wine bug was that Steam was trying to keep installing MSVC
forever in a loop, at each startup attempt.