buildsystem: no transitive linkage for internally used libraries
if the linking is declared PUBLIC, e.g. all the python modules that
link against libopenage would link to all of libopenage's dependencies
again. when one of those, e.g. libnyan.so, is updated, the module will be
linked again. this means if e.g. libnyan is updated, all python modules
are relinked. this commit removes this transitive linking, and python
modules won't be rebuilt any more.